PI with Chudnovsky method
a simple  implementation using yabIntType, but without  Binary Splitting .  Not as fast as a Python implementation, but yabIntType is for simplicity and portability, not speed.
Calculates around 2000 digits of PI for R =20 in 20minutes on my old PC,  most of the time is spent doing an  Integer Squaretroot  with class Newton-Rapshon iteration. 
This is mostly as a test of yabIntType, but also a bit of fun mathematics.
Also some vague ideas about faster squareroot calculations.